requirements.txt
# pip3 install -U -r requirements.txt
numpy == 1.17.3
scipy == 1.4.1
Pillow == 6.2.1
opencv-python >= 4.1.0
argparse == 1.4.0